Create LICENSE
Under this license, any unauthorized use of personal details or information provided within this repository is strictly prohibited. All users must adhere to the specified conditions to ensure the privacy and security of the information. Violations of these terms may result in legal action. This license will be utilized for all my projects. Thanks for understanding.